+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## ScanPlus [1.0.3] - 2024-03-22
+
+### Added
+
+- Added compatibility with `TerminalFormatter`. ([#1](https://github.com/AidanTweedy/scanplus/issues/1))
+
+### Changed
+
+- Updated README
+
+### Removed
+
+- None
+
 ## ScanPlus [1.0.2] - 2024-03-21
 
 ### Added